One script to rule them all...
In the land of Mordor... No wait...
Use a single Python script for each service. This reduces duplicate
code and should be easier to maintain long term. Adds argparse support
to determine tool to be used. We can add more optional arguments in the
future (debug, output file names, etc).